-Updated AIManager with new RemoveFromDarknessList function.  To remove AI properly from the list and allow spawning of more AI during different stages of battle. -Adjusting the intensity of battle on stage 1, stage 2, and stage 3 of battle. -Fixing DarkBoss attack animation. Messed up because somehow deleted prefab -pushing scene changes to song2

+++ Assets/Scripts/Monsters/EnemySpawner.cs
@@ -11,7 +11,7 @@ public class EnemySpawner : MonoBehaviour
     [SerializeField]
     Darkness darknessEnemy;
     [SerializeField]
-    public int spawnRate = 10; //The greater the slower, 1 every x seconds
+    public float spawnRate = 10; //The greater the slower, 1 every x seconds
     [SerializeField]
     Transform[] spawnLocs;
